re-arrange and explain
I've put the line for downloading the data at the top and deleted the
appendix because it was not explained that more data had to be
downloaded. I would prefer to do it like this in "getting started" as
well. If there's a reason for putting the download in an appendix then
ok, but there then needs to be a line here telling the student to do it.
I have not made any changes to the lines beginning with underscore, eg:
_ = s3.plot_bss_loadings()
I don't know what this means. I always just type "s3.plot..." Is it
explained somewhere? (Sorry if I've missed it).

By the way if it is judged too soon to the workshop to start messing
around with these demos just now, then let me know and I'll leave it for
now.